requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.0
gunicorn==21.2.0
playwright==1.40.0
